Add Optional Cluster UID Validation (#4)
This PR adds an optional `uid` field that can be set in a kubeapply config. If set, kubeapply will enforce that this uid matches the `kube-system` namespace uid of the cluster you are diffing/applying.

In general, we will already catch if you are running kubeapply on the wrong cluster by grepping for the cluster name in the kubeconfig, but if your cluster has the same name in different environments, then that solution won't detect it. If you have clusters of the same name in different environments, you'll likely want to set this UID.