feat: Upgrade to Kubernetes 1.21
Kubernetes dependencies are upgraded to the latest 1.21 release.

A module replacement for github.com/golang/protobuf to facilitate
skycfg functionality has been commented out as well, in the hopes that
things are working almost two years on. Unfortunately, the root skycfg
issue remains open. In the meantime, Segment has ceased experimenting
with it.